+++ lib/falcon/command/proxy.rb
@@ -43,7 +43,7 @@ module Falcon
 			end
 			
 			def configuration
-				Configuration.for(
+				Async::Service::Configuration.for(
 					self.environment(environments: super.environments)
 				)
 			end