+++ lib/falcon/environment.rb
@@ -5,9 +5,6 @@
 
 module Falcon
 	# Pre-defined environments for hosting web applications.
-	#
-	# See {Configuration::Loader#load} for more details.
 	module Environment
-		LEGACY_ENVIRONMENTS = {}
 	end
 end